requirements.txt
packaging==24.0
typing_extensions==4.10.0
typing-inspect==0.9.0
jsonlines==4.0.0
joblib==1.3.2
tqdm==4.66.3
python-dotenv==1.0.1
wheel==0.44.0
# data
pandas==2.2.2
matplotlib==3.9.1
seaborn==0.13.2
# deep learning
torch==2.4.0
transformers==4.43.4
trl==0.9.6
tokenizers==0.19.1
accelerate==0.32.0
peft==0.12.0
datasets==2.20.0
deepspeed==0.15.0
bitsandbytes==0.43.3
sentencepiece==0.2.0
flash-attn==2.6.3
# openai
tiktoken==0.7.0
openai==1.39.0
# observability
wandb==0.17.5
loguru==0.7.2
# notebook
ipykernel==6.29.5
ipywidgets==8.1.3
ipython==8.16.1 # for wandb
# formatting
black==24.8.0
isort==5.13.2
